separated corner and model activities
